代码整理ok

+++ routers/views/admin.go
@@ -9,73 +9,147 @@ import (
 	"github.com/likiiiiii/foxpot_backend/utils"
 )
 
-// GetAdminIndex 管理员首页
+// GetAdminIndex 管理员首页 默认攻击地图
 func GetAdminIndex(c *gin.Context) {
 	session := sessions.Default(c)
 	userID, _ := session.Get(utils.Config.Session.Key).(uint)
 	user, _ := models.GetUserByID(userID)
-	c.HTML(http.StatusOK, "admin/index.html", gin.H{
-		"code":    200,
-		"message": "欢迎欢迎",
-		"title":   "首页",
-		"User":    user,
+	c.HTML(http.StatusOK, "admin/attackmap", gin.H{
+		"title": "攻击地图",
+		"user":  user,
 	})
 }
 
-// GetAdminProfile 管理员个人中心
+// GetAdminProfile 个人中心
 func GetAdminProfile(c *gin.Context) {
 	session := sessions.Default(c)
 	userID, _ := session.Get(utils.Config.Session.Key).(uint)
 	user, _ := models.GetUserByID(userID)
-	c.HTML(http.StatusOK, "admin/profile.html", gin.H{
-		"code":    200,
-		"message": "管理员个人中心",
-		"title":   "个人中心",
-		"User":    user,
+	c.HTML(http.StatusOK, "admin/profile", gin.H{
+		"title": "个人中心",
+		"user":  user,
 	})
+}
+
+// GetAdminUpdateProfile 修改个人资料
+func GetAdminUpdateProfile(c *gin.Context) {
+	session := sessions.Default(c)
+	userID, _ := session.Get(utils.Config.Session.Key).(uint)
+	user, _ := models.GetUserByID(userID)
+	c.HTML(http.StatusOK, "admin/updateprofile", gin.H{
+		"title": "修改个人资料",
+		"user":  user,
+	})
+}
+
+// PostAdminUpdateProfile 提交个人资料
+func PostAdminUpdateProfile(c *gin.Context) {
+	session := sessions.Default(c)
+	userID, _ := session.Get(utils.Config.Session.Key).(uint)
+	user, _ := models.GetUserByID(userID)
+	// TODO
+	// 管理员用户修改个人资料也就更改密码 邮箱 手机号得了
+	password := c.PostForm("password")
+	confirmPassword := c.PostForm("confirm")
+	email := c.PostForm("email")
+	phone := c.PostForm("phone")
+	if password != confirmPassword {
+		c.HTML(http.StatusBadRequest, "admin/updateprofile", gin.H{
+			"message": "密码不一致",
+			"title":   "修改个人资料",
+			"user":    user,
+		})
+		return
+	}
+	if password != "" {
+		password, _ = utils.HashPassword(password)
+		user.HashedPassword = password
+	}
+	user.Email = email
+	user.Phone = phone
+	err := models.UpdateUser(user)
+	if err != nil {
+		c.HTML(http.StatusOK, "admin/updateprofile", gin.H{
+			"message": "更新失败",
+			"title":   "修改个人资料",
+			"user":    user,
+		})
+	} else {
+		c.HTML(http.StatusOK, "admin/updateprofile", gin.H{
+			"message": "更新成功",
+			"title":   "修改个人资料",
+			"user":    user,
+		})
+	}
 }
 
 // GetAdminUserManage 用户管理
 func GetAdminUserManage(c *gin.Context) {
+	session := sessions.Default(c)
+	userID, _ := session.Get(utils.Config.Session.Key).(uint)
+	user, _ := models.GetUserByID(userID)
 	users, _ := models.GetAllUsers()
-	c.HTML(http.StatusOK, "admin/usermanage.html", gin.H{
-		"code":    200,
-		"message": "用户列表",
-		"title":   "用户管理",
-		"Users":   users,
+	c.HTML(http.StatusOK, "admin/usermanage", gin.H{
+		"title": "用户管理",
+		"user":  user,
+		"users": users,
 	})
 }
 
 // GetAdminCreateUser 新建用户
 func GetAdminCreateUser(c *gin.Context) {
-	c.HTML(http.StatusOK, "admin/createuser.html", gin.H{
-		"code":    200,
-		"message": "",
-		"title":   "新建用户",
+	session := sessions.Default(c)
+	userID, _ := session.Get(utils.Config.Session.Key).(uint)
+	user, _ := models.GetUserByID(userID)
+	c.HTML(http.StatusOK, "admin/createuser", gin.H{
+		"title": "新建用户",
+		"user":  user,
 	})
 }
 
 // PostAdminCreateUser 新建用户
 func PostAdminCreateUser(c *gin.Context) {
+	session := sessions.Default(c)
+	userID, _ := session.Get(utils.Config.Session.Key).(uint)
+	user, _ := models.GetUserByID(userID)
 	username := c.PostForm("username")
 	password := c.PostForm("password")
-	role := c.PostForm("role")
+	confirmPassword := c.PostForm("confirm")
+	role := models.Role2Uint(c.PostForm("role"))
 	email := c.PostForm("email")
 	phone := c.PostForm("phone")
-	if username == "" || password == "" || role == "" {
-		c.HTML(http.StatusBadRequest, "admin/createuser.html", gin.H{
-			"code":    400,
-			"message": "用户名和密码不能为空",
+	if username == "" || password == "" {
+		c.HTML(http.StatusBadRequest, "admin/createuser", gin.H{
+			"message": "用户名/密码不能为空",
 			"title":   "新建用户",
+			"user":    user,
 		})
 		return
 	}
-	// check if already existed
-	if _, err := models.GetUserByUsername(username); err == nil {
-		c.HTML(http.StatusBadRequest, "admin/createuser.html", gin.H{
-			"code":    400,
-			"message": "用户已存在",
+	// check if password != confirmPassword
+	if password != confirmPassword {
+		c.HTML(http.StatusBadRequest, "admin/createuser", gin.H{
+			"message": "密码不一致",
 			"title":   "新建用户",
+			"user":    user,
+		})
+		return
+	}
+	// check if superadmin
+	if user.Role != 0 && role == 1 {
+		c.HTML(http.StatusBadRequest, "admin/createuser", gin.H{
+			"message": "只有超级管理员能够新建管理员",
+			"title":   "新建用户",
+			"user":    user,
+		})
+		return
+	}
+	// check if user already existed
+	if _, err := models.GetUserByUsername(username); err == nil {
+		c.HTML(http.StatusBadRequest, "admin/createuser", gin.H{
+			"message": "用户名已存在",
+			"title":   "新建用户",
+			"user":    user,
 		})
 		return
 	}
@@ -88,27 +162,36 @@ func PostAdminCreateUser(c *gin.Context) {
 		Phone:          phone,
 	})
 	if err != nil {
-		c.HTML(http.StatusOK, "admin/createuser.html", gin.H{
-			"code":    500,
+		c.HTML(http.StatusOK, "admin/createuser", gin.H{
 			"message": "新建失败",
 			"title":   "新建用户",
+			"user":    user,
 		})
 	} else {
-		c.HTML(http.StatusOK, "admin/createuser.html", gin.H{
-			"code":    200,
+		c.HTML(http.StatusOK, "admin/createuser", gin.H{
 			"message": "新建成功",
 			"title":   "新建用户",
+			"user":    user,
 		})
 	}
 }
 
 // PostAdminDeleteUser 删除用户
 func PostAdminDeleteUser(c *gin.Context) {
+	session := sessions.Default(c)
+	userID, _ := session.Get(utils.Config.Session.Key).(uint)
+	user, _ := models.GetUserByID(userID)
 	var (
 		usernames []string
 		deleted   []string
 	)
 	c.ShouldBind(&usernames)
+	if user.Role != 0 && models.HasAnyAdminByUsernames(usernames) {
+		c.JSON(http.StatusOK, gin.H{
+			"message": "只有超级管理员能够删除管理员",
+		})
+		return
+	}
 	for _, username := range usernames {
 		// TODO 做成事务  错误回滚
 		if err := models.DeleteUserByUsername(username); err == nil {
@@ -116,7 +199,6 @@ func PostAdminDeleteUser(c *gin.Context) {
 		}
 	}
 	c.JSON(http.StatusOK, gin.H{
-		"code":    200,
 		"message": "删除成功",
 		"data":    deleted,
 	})
@@ -124,38 +206,153 @@ func PostAdminDeleteUser(c *gin.Context) {
 
 // GetAdminUpdateUser 更新用户信息
 func GetAdminUpdateUser(c *gin.Context) {
-	// TODO 参数判定下 这样不严格 防乱搞
-	username := c.DefaultQuery("username", "none")
-	user, _ := models.GetUserByUsername(username)
-	c.HTML(http.StatusOK, "admin/updateuser.html", gin.H{
-		"code":    200,
-		"message": "",
-		"title":   "更新用户信息",
-		"User":    user,
+	session := sessions.Default(c)
+	userID, _ := session.Get(utils.Config.Session.Key).(uint)
+	user, _ := models.GetUserByID(userID)
+	// TODO
+	// 检查参数防乱搞
+	username := c.Param("username")
+	editUser, _ := models.GetUserByUsername(username)
+	if user.Role != 0 && editUser.IsAdmin() {
+		users, _ := models.GetAllUsers()
+		c.HTML(http.StatusOK, "admin/usermanage", gin.H{
+			"message": "只有超级管理员能够编辑管理员",
+			"title":   "用户管理",
+			"user":    user,
+			"users":   users,
+		})
+		return
+	}
+	c.HTML(http.StatusOK, "admin/updateuser", gin.H{
+		"title":    "编辑用户信息",
+		"user":     user,
+		"editUser": editUser,
 	})
 }
 
 // PostAdminUpdateUser 更新用户信息
 func PostAdminUpdateUser(c *gin.Context) {
+	session := sessions.Default(c)
+	userID, _ := session.Get(utils.Config.Session.Key).(uint)
+	user, _ := models.GetUserByID(userID)
+	// TODO
+	// 检查参数防乱搞
 	username := c.PostForm("username")
-	user, _ := models.GetUserByUsername(username)
-	user.Role = c.PostForm("role")
-	user.Email = c.PostForm("email")
-	user.Phone = c.PostForm("phone")
-	err := models.UpdateUser(user)
-	if err != nil {
-		c.HTML(http.StatusOK, "admin/updateuser.html", gin.H{
-			"code":    500,
-			"message": "更新失败",
-			"title":   "更新用户信息",
-			"User":    user,
+	password := c.PostForm("password")
+	confirmPassword := c.PostForm("confirm")
+	role := models.Role2Uint(c.PostForm("role"))
+	email := c.PostForm("email")
+	phone := c.PostForm("phone")
+	editUser, _ := models.GetUserByUsername(username)
+	// check if password != confirmPassword
+	if password != confirmPassword {
+		c.HTML(http.StatusBadRequest, "admin/updateuser", gin.H{
+			"message":  "密码不一致",
+			"title":    "新建用户",
+			"user":     user,
+			"editUser": editUser,
 		})
-	} else {
-		c.HTML(http.StatusOK, "admin/updateuser.html", gin.H{
-			"code":    200,
-			"message": "更新成功",
-			"title":   "更新用户信息",
-			"User":    user,
+		return
+	}
+	// check if superadmin
+	if user.Role != 0 && role < editUser.Role {
+		c.HTML(http.StatusBadRequest, "admin/updateuser", gin.H{
+			"message":  "只有超级管理员能够提升权限",
+			"title":    "编辑用户信息",
+			"user":     user,
+			"editUser": editUser,
 		})
 	}
+	editUser.Username = username
+	if password != "" {
+		password, _ = utils.HashPassword(password)
+		editUser.HashedPassword = password
+	}
+	editUser.Role = role
+	editUser.Email = email
+	editUser.Phone = phone
+	err := models.UpdateUser(editUser)
+	if err != nil {
+		c.HTML(http.StatusOK, "admin/updateuser", gin.H{
+			"message":  "更新失败",
+			"title":    "编辑用户信息",
+			"user":     user,
+			"editUser": editUser,
+		})
+	} else {
+		c.HTML(http.StatusOK, "admin/updateuser", gin.H{
+			"message":  "更新成功",
+			"title":    "编辑用户信息",
+			"user":     user,
+			"editUser": editUser,
+		})
+	}
+}
+
+// 第三方组件
+
+// GetAdminESHead ESHead
+func GetAdminESHead(c *gin.Context) {
+	session := sessions.Default(c)
+	userID, _ := session.Get(utils.Config.Session.Key).(uint)
+	user, _ := models.GetUserByID(userID)
+	c.HTML(http.StatusOK, "admin/eshead", gin.H{
+		"title": "ES Head",
+		"user":  user,
+	})
+}
+
+// GetAdminKibanaDiscover KibanaDiscover
+func GetAdminKibanaDiscover(c *gin.Context) {
+	session := sessions.Default(c)
+	userID, _ := session.Get(utils.Config.Session.Key).(uint)
+	user, _ := models.GetUserByID(userID)
+	c.HTML(http.StatusOK, "admin/discover", gin.H{
+		"title": "告警详情",
+		"user":  user,
+	})
+}
+
+// GetAdminKibanaDashboard KibanaDashboard
+func GetAdminKibanaDashboard(c *gin.Context) {
+	session := sessions.Default(c)
+	userID, _ := session.Get(utils.Config.Session.Key).(uint)
+	user, _ := models.GetUserByID(userID)
+	c.HTML(http.StatusOK, "admin/dashborad.html", gin.H{
+		"title": "仪表板",
+		"user":  user,
+	})
+}
+
+// GetAdminCockpitDocker Cockpit容器状态监控
+func GetAdminCockpitDocker(c *gin.Context) {
+	session := sessions.Default(c)
+	userID, _ := session.Get(utils.Config.Session.Key).(uint)
+	user, _ := models.GetUserByID(userID)
+	c.HTML(http.StatusOK, "admin/docker", gin.H{
+		"title": "系统状态",
+		"User":  user,
+	})
+}
+
+// GetAdminCockpitSystem Cockpit系统状态监控
+func GetAdminCockpitSystem(c *gin.Context) {
+	session := sessions.Default(c)
+	userID, _ := session.Get(utils.Config.Session.Key).(uint)
+	user, _ := models.GetUserByID(userID)
+	c.HTML(http.StatusOK, "admin/system", gin.H{
+		"title": "系统状态",
+		"User":  user,
+	})
+}
+
+// GetAdminCockpitTerminal CockpitWeb终端
+func GetAdminCockpitTerminal(c *gin.Context) {
+	session := sessions.Default(c)
+	userID, _ := session.Get(utils.Config.Session.Key).(uint)
+	user, _ := models.GetUserByID(userID)
+	c.HTML(http.StatusOK, "admin/terminal", gin.H{
+		"title": "Web终端",
+		"User":  user,
+	})
 }
